+++ lib/db/model/folders.js
@@ -137,43 +137,42 @@ exports.updateStatus = function (id, status) {
 
 
 
-var clearDatabase = schedule.scheduleJob({
-    hour: 21,
-    minute: 52
-}, function () {
-    exports.deleteJunk = function () {
-        var deferred = Q.defer();
-        Folders.find({
-            status: false
-        }, function (err, filesStatusFalse) {
+// var clearDatabase = schedule.scheduleJob({
+//     hour: 9,
+//     minute: 0
+// }, function () {
+//     exports.deleteJunk = function () {
+//         var deferred = Q.defer();
+//         Folders.find({
+//             status: false
+//         }, function (err, filesStatusFalse) {
 
-            var currentDate = new Date();
+//             var currentDate = new Date();
 
-            filesStatusFalse.forEach(function (files) {
+//             filesStatusFalse.forEach(function (files) {
 
-                var dbDate = files.updated;
+//                 var dbDate = files.updated;
 
-                var diff = (currentDate - dbDate) / 60e3;
-                //60e3 minute
-                ///1e3 sekunde
+//                 var diff = (currentDate - dbDate) / 60e3;
+//                 //60e3 minute
+//                 ///1e3 sekunde
 
-                console.log('diff', diff);
+//                 console.log('diff', diff);
 
-                if (diff < 10) {
-                    Folders.remove({
-                        _id: files._id
-                    }, function (err, filesForDelete) {
-                        if (err) {
-                            logger.error('ERROR DB - Deleting Files with status = false', err);
-                            return deferred.reject(error('MONGO_ERROR'));
-                        }
-                        deferred.resolve(filesForDelete);
-                    });
+//                 if (diff < 10) {
+//                     Folders.remove({
+//                         _id: files._id
+//                     }, function (err, filesForDelete) {
+//                         if (err) {
+//                             logger.error('ERROR DB - Deleting Files with status = false', err);
+//                             return deferred.reject(error('MONGO_ERROR'));
+//                         }
+//                         deferred.resolve(filesForDelete);
+//                     });
 
-                }
-            });
-        });
-
-        return deferred.promise;
-    }();
-});
+//                 }
+//             });
+//         });
+//         return deferred.promise;
+//     }();
+// });
